+++ internal/keyseq/ahocorasick.go
