+++ internal/keyseq/ahocorasick_test.go
