+++ internal/keyseq/keys_test.go
