+++ internal/keyseq/ternary.go
