+++ internal/keyseq/ternary_test.go
