+++ internal/keyseq/trie.go
