+++ internal/util/homedir_darwin.go
@@ -1,11 +1,11 @@
-package peco
+package util
 
 import (
 	"fmt"
 	"os"
 )
 
-func homedir() (string, error) {
+func Homedir() (string, error) {
 	home := os.Getenv("HOME")
 	if home == "" {
 		return "", fmt.Errorf("error: Environment variable HOME not set")
