+++ internal/util/homedir_posix.go
@@ -1,13 +1,13 @@
 // +build !darwin,!windows
 
-package peco
+package util
 
 import (
 	"fmt"
 	"os"
 )
 
-func homedir() (string, error) {
+func Homedir() (string, error) {
 	home := os.Getenv("HOME")
 	if home == "" {
 		return "", fmt.Errorf("error: Environment variable HOME not set")
