+++ internal/util/tty_bsd.go
@@ -1,6 +1,6 @@
 // +build darwin freebsd openbsd netbsd dragonfly
 
-package peco
+package util
 
 import (
 	"syscall"
