+++ internal/util/tty_windows.go
@@ -1,4 +1,4 @@
-package peco
+package util
 
 import (
 	"os"
