+++ internal/util/util.go
@@ -1,0 +1,23 @@
+package util
+
+import (
+	"regexp"
+	"unicode"
+)
+
+func ContainsUpper(query string) bool {
+	for _, c := range query {
+		if unicode.IsUpper(c) {
+			return true
+		}
+	}
+	return false
+}
+
+// Global var used to strips ansi sequences
+var reANSIEscapeChars = regexp.MustCompile("\x1B\\[(?:[0-9]{1,2}(?:;[0-9]{1,2})?)*[a-zA-Z]")
+
+// Function who strips ansi sequences
+func StripANSISequence(s string) string {
+	return reANSIEscapeChars.ReplaceAllString(s, "")
+}
