+++ layout_windows.go
@@ -1,0 +1,8 @@
+package peco
+
+func init() {
+	// This is the value we pass to anchor offset when we're running
+	// on windows platform
+	extraOffset = 1
+}
+
