+++ line.go
@@ -1,15 +1,11 @@
 package peco
 
 import (
-	"regexp"
 	"strings"
 
 	"github.com/google/btree"
+	"github.com/peco/peco/internal/util"
 )
-
-type idGen struct {
-	genCh chan uint64
-}
 
 func newIDGen() *idGen {
 	ch := make(chan uint64)
@@ -29,55 +25,6 @@ func newIDGen() *idGen {
 
 func (ig *idGen) create() uint64 {
 	return <-ig.genCh
-}
-
-// Global var used to strips ansi sequences
-var reANSIEscapeChars = regexp.MustCompile("\x1B\\[(?:[0-9]{1,2}(?:;[0-9]{1,2})?)*[a-zA-Z]")
-
-// Function who strips ansi sequences
-func stripANSISequence(s string) string {
-	return reANSIEscapeChars.ReplaceAllString(s, "")
-}
-
-// Line represents each of the line that peco uses to display
-// and match against queries.
-type Line interface {
-	btree.Item
-
-	ID() uint64
-
-	// Buffer returns the raw buffer
-	Buffer() string
-
-	// DisplayString returns the string to be displayed. This means if you have
-	// a null separator, the contents after the separator are not included
-	// in this string
-	DisplayString() string
-
-	// Indices return the matched portion(s) of a string after filtering.
-	// Note that while Indices may return nil, that just means that there are
-	// no substrings to be highlighted. It doesn't mean there were no matches
-	Indices() [][]int
-
-	// Output returns the string to be display as peco finishes up doing its
-	// thing. This means if you have null separator, the contents before the
-	// separator are not included in this string
-	Output() string
-
-	// IsDirty returns true if this line should be forcefully redrawn
-	IsDirty() bool
-
-	// SetDirty sets the dirty flag on or off
-	SetDirty(bool)
-}
-
-// RawLine is the input line as sent to peco, before filtering and what not.
-type RawLine struct {
-	id            uint64
-	buf           string
-	sepLoc        int
-	displayString string
-	dirty         bool
 }
 
 var idGenerator = newIDGen()
@@ -138,9 +85,9 @@ func (rl RawLine) DisplayString() string {
 	}
 
 	if i := rl.sepLoc; i > -1 {
-		rl.displayString = stripANSISequence(rl.buf[:i])
+		rl.displayString = util.StripANSISequence(rl.buf[:i])
 	} else {
-		rl.displayString = stripANSISequence(rl.buf)
+		rl.displayString = util.StripANSISequence(rl.buf)
 	}
 	return rl.displayString
 }
@@ -157,12 +104,6 @@ func (rl RawLine) Output() string {
 // returns nil
 func (rl RawLine) Indices() [][]int {
 	return nil
-}
-
-// MatchedLine contains the indices to the matches
-type MatchedLine struct {
-	Line
-	indices [][]int
 }
 
 // NewMatchedLine creates a new MatchedLine
