+++ screen_posix.go
@@ -1,0 +1,7 @@
+// +build !windows
+
+package peco
+
+func (t Termbox) PostInit() error {
+	return nil
+}
