+++ screen_windows.go
@@ -1,0 +1,10 @@
+package peco
+
+import "github.com/nsf/termbox-go"
+
+func (t Termbox) PostInit() error {
+	// Windows handle Esc/Alt self
+	termbox.SetInputMode(termbox.InputEsc | termbox.InputAlt)
+
+	return nil
+}
