+++ selection.go
@@ -2,11 +2,6 @@ package peco
 
 import "github.com/google/btree"
 
-// Selection stores the line ids that were selected by the user.
-// The contents of the Selection is always sorted from smallest to
-// largest line ID
-type Selection struct{ *btree.BTree }
-
 // NewSelection creates a new empty Selection
 func NewSelection() *Selection {
 	return &Selection{btree.New(32)}
